doc2qa/requirements.txt
streamlit 
python-docx 
pdfplumber 
pytesseract 
pillow 
python-dotenv 
groq